+++ packages/blocks/core/js/third-party/bootstraps.js
@@ -6,3 +6,4 @@ export * from './blocksy-about-me/bootstrap';
 export * from './blocksy-contact-info/bootstrap';
 export * from './blocksy-socials/bootstrap';
 export * from './blocksy-share-box/bootstrap';
+export * from './blocksy-search/bootstrap';